alx_travel_app_0x02/requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
requests
celery
redis
django-celery-beat
django-celery-results
pymysql
django-environ

